separate server and app
